Revert "Add ThankYouPage model"
This reverts commit 067f726fb0959db668ae9d8624b97bb30d56df96.

I have finally spent some time shaping how I want this form feature
to behave.

Since I am creating this prototype for a modern headless setup, I also
decided that it is only consistent to use JS for the form submissions.
This also makes the whole concept a bit clearer and more consistent.

I will use JS on the frontend to submit the form instead of directly
sending the form with an HTTP request. This allows me to send form
validation errors as JSON back to the frontend. JS can then show
the error in the form.

Since I am now sending the data with JS to the backend, I also need to
handle successful submissions with JS. I could still use JS to redirect
to a thank you page. But, I have decided against that for now. I could
simply use JS to show the success status in the form itself. This also
solves the issue of how the user gets back from the Thank You page to
the normal website. The user will simply stay on the page where the form
is shown and will see a success message.

This all feels a little more sensible since I am not planning on using
the Form pages (although defined like pages) as stand alone pages.
Rather, I want to include the form in other pages with some more
context.

For now, I am removing the thank you pages and rather go with a thank
you message that is defined on the form itself.