Add test and form template for GET
The forms, while defined as pages, are not intended to be used as standalone pages
by themselves. Rather, they will be included in another pages that gives context
to the form. This allows the forms to be reused and should make the mental context
in which a vistor may use the form more sensible.

Not useing the forms as pages creates the situation that there are still are the links
to view the live or preview version of the form in the admin. To inform the admin users
about whay they are not seeing the form, the template is added. For now it only shows a
message to see the pages in which the form is used.

It could be a nice addition to actually show a list of links to pages where the form is
actually used.