Test POST invalid form data leads to error resp
This is just the test. Other then the current implementation, I want an HTTP error response when the
submitted data contains validation errors.

The current implementation still leds the original `serve` method handle the POST
request. The default behaviour is to render the form error into the template. Because the template
is successfully sent, this comes with a 200 status code.

I do not want a template response, but a JSON response that indicates an error.