Remove reliance on super().serve from GET handling
Since I want the serve method to behave much differently than the default behavior,
it only creates mental overhead and constant need to look at the original `serve` implementation
when I keep invoking it in the GET and POST handling.

This commit remove the need to call the `serve` method of the super class in the GET handling.
It is now much more transparent what is happening on when a GET request is received,
since it does not depent on the super classes implementation.