Make test for saving/rejecting data more explicit
Before, I was only checking that when processing successful form submission that the number of
submissions was increased (or not incase of a spam submission)

Now, I am testing for actually having 0 submissions in the beginning of the test and 1 after the
submission is processed. This makes the test less flexible, but more explicit. If I am creating
additional tests in the future, this change also makes sure that the tests do not have unintended
side effects that persist submissions over test cases.

If I create tests that do require this kind of persistence over multiple test cases, these side effects
have to be explicitly accepted as test results.