Test valid form submission sends email to admin
This adds a test that a valid form submission creates an email in the
`mailoutbox`. This is a fixture provided by `pytest-django`. The actual
sending still needs to be checked, but should be a matter of config
rather than code.

Next tests should make sure that invalid of spam submissions do not
trigger emails.