Rename page for grapple issue and add relation
For some reason, grapple was stipping the "Content" from the "ContentPage" model.
This lead to a duplicate schema definition. To get around this, I renamed it to
the fantastic name StuffPage, which works.

Also, I discorvered this while I was trying to test the created relation from the
content/stuff page to the contact form. So that is implemented with this commit too.
;