Specifiy errors in error response
Before, the error response that was sent on invalid form data did immediately
contain the field names and then the error codes and error messages.

Looking only at the data, you need to read the messages to understand
that these are error messages for the different fields.

I added an additional nesting level that labels the error data as `errors`.
This should be more obvious to understand.